+++ src/main/java/com/robertx22/age_of_exile/mmorpg/init/CommonInit.java
@@ -14,6 +14,7 @@ import com.robertx22.age_of_exile.mmorpg.registers.common.C2SPacketRegister;
 import com.robertx22.age_of_exile.mmorpg.registers.common.ConfigRegister;
 import com.robertx22.age_of_exile.mmorpg.registers.common.MobAttributes;
 import com.robertx22.age_of_exile.mmorpg.registers.common.ModCriteria;
+import com.robertx22.age_of_exile.uncommon.utilityclasses.LevelUtils;
 import me.sargunvohra.mcmods.autoconfig1u.AutoConfig;
 import me.sargunvohra.mcmods.autoconfig1u.serializer.GsonConfigSerializer;
 import net.fabricmc.api.ModInitializer;
@@ -81,7 +82,23 @@ public class CommonInit implements ModInitializer {
 
         }
 
+        testLevelCurve();
+
         uncapHealth();
+
+    }
+
+    public static void testLevelCurve() {
+        for (int i = 1; i < ModConfig.get().Server.MAX_LEVEL + 1; i++) {
+
+            int needed = LevelUtils.getExpRequiredForLevel(i);
+            int basepermob = LevelUtils.getBaseExpMobReward(i);
+            int killsneeded = needed / basepermob;
+            System.out.print("\nExp needed for lvl: " + i + " is " + needed);
+            System.out.print("\nBase mob xp reward for lvl: " + i + " is " + basepermob);
+            System.out.print("\nKills needed for level: " + i + " is " + killsneeded);
+            System.out.print("\n");
+        }
     }
 
     static void uncapHealth() {