+++ src/main/java/com/robertx22/age_of_exile/uncommon/utilityclasses/OnScreenMessageUtils.java
@@ -28,4 +28,9 @@ public class OnScreenMessageUtils {
 
     }
 
+    public static void sendMessage(ServerPlayerEntity p, MutableText title, TitleS2CPacket.Action act) {
+        p.networkHandler.sendPacket(new TitleS2CPacket(act, title
+            , 5, 15, 8));
+    }
+
 }